refactor: avoid events for uploading
